# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Game Evolution",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8080",
    "basePath": "/",
    "paths": {
        "/api/cards": {
            "get": {
                "description": "Возвращает всех доуступных карты с опциональной фильтрацией по ФИО",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "карты"
                ],
                "summary": "Получить всех карты",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ФИО для фильтрации",
                        "name": "fio",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.GetAllCardsResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Добавить нового карты",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "карты"
                ],
                "summary": "Добавить карты",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Изображение карты",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "ФИО",
                        "name": "fio",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Почта",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Возраст",
                        "name": "age",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Адрес",
                        "name": "adress",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/cards/{id}": {
            "get": {
                "description": "Возвращает более подробную информацию об одной карте",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "карты"
                ],
                "summary": "Получить одного карты",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id карты",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ds.Card"
                        }
                    }
                }
            },
            "put": {
                "description": "Изменить данные полей о карте",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "карты"
                ],
                "summary": "Изменить карты",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор карты",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ФИО",
                        "name": "fio",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Почта",
                        "name": "email",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "Возраст",
                        "name": "age",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Изображение карты",
                        "name": "image",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Адрес",
                        "name": "adress",
                        "in": "formData"
                    }
                ],
                "responses": {}
            },
            "delete": {
                "description": "Удаляет карты по id",
                "tags": [
                    "карты"
                ],
                "summary": "Удалить карты",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id карты",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/cards/{id}/add_to_turn": {
            "post": {
                "description": "Добавить выбранную карту в черновик хода",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "карты"
                ],
                "summary": "Добавить в ход",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id карты",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schemes.AddToTurnResp"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ds.Card": {
            "type": "object",
            "required": [
                "description",
                "name",
                "needfood",
                "type"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 200
                },
                "image_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 50
                },
                "needfood": {
                    "type": "integer"
                },
                "type": {
                    "type": "string",
                    "maxLength": 50
                },
                "uuid": {
                    "type": "string"
                }
            }
        },
        "schemes.AddToTurnResp": {
            "type": "object",
            "properties": {
                "card_count": {
                    "type": "integer"
                }
            }
        },
...
        "schemes.GetAllCardsResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ds.Card"
                    }
                },
                "draft_turn": {
                    "$ref": "#/definitions/schemes.TurnShort"
                }
            }
        },
...
        "schemes.TurnShort": {
            "type": "object",
            "properties": {
                "card_count": {
                    "type": "integer"
                },
                "uuid": {
                    "type": "string"
                }
            }
        }
    }
}